+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - define environment variables
     - set the current working directory
     - print more error details if a test fails
+    - add field validation
 
 # v0.1.0